requirements.txt
Pillow == 10.0.0
fastapi == 0.115.6
uvicorn == 0.34.0
Jinja2 == 3.1.2
python-multipart==0.0.20